Split schemas from models
