+++ app/schemas/ingestion.py
@@ -1,0 +1,117 @@
+from __future__ import annotations
+
+from datetime import datetime
+from typing import Any, Literal
+
+from pydantic import BaseModel, Field, field_validator
+
+from .domain import CSVImportType, CTEType, DestinationMode, RegEngineEvent
+from .simulation import DeliveryConfig
+
+
+class IngestPayload(BaseModel):
+    source: str = "codex-simulator"
+    events: list[RegEngineEvent]
+
+
+class IngestResponseEvent(BaseModel):
+    traceability_lot_code: str
+    cte_type: CTEType
+    status: Literal["accepted", "rejected"]
+    event_id: str
+    sha256_hash: str
+    chain_hash: str
+
+
+class MockIngestResponse(BaseModel):
+    accepted: int
+    rejected: int
+    total: int
+    events: list[IngestResponseEvent]
+    ingestion_timestamp: datetime
+
+
+class DeliveryRetryRequest(BaseModel):
+    record_ids: list[str] | None = None
+    limit: int = 50
+    source: str | None = None
+    delivery: DeliveryConfig | None = None
+
+    @field_validator("limit")
+    @classmethod
+    def validate_limit(cls, value: int) -> int:
+        if value < 1 or value > 500:
+            raise ValueError("limit must be between 1 and 500")
+        return value
+
+
+class DeliveryRetryResponse(BaseModel):
+    status: Literal["empty", "posted", "partial", "failed", "skipped"]
+    requested: int
+    retryable: int
+    attempted: int
+    posted: int
+    failed: int
+    skipped: int
+    delivery_mode: DestinationMode
+    record_ids: list[str]
+    responses: list[dict[str, Any]] = Field(default_factory=list)
+    error: str | None = None
+
+
+class ReplayRequest(BaseModel):
+    persist_path: str | None = None
+    source: str | None = None
+    delivery: DeliveryConfig | None = None
+
+
+class ReplayResponse(BaseModel):
+    status: Literal["empty", "posted", "rebuilt", "failed"]
+    read: int
+    replayed: int
+    posted: int
+    failed: int
+    source: str
+    persist_path: str
+    delivery_mode: DestinationMode
+    delivery_attempts: int = 0
+    response: dict[str, Any] | None = None
+    error: str | None = None
+
+
+class CSVImportRequest(BaseModel):
+    import_type: CSVImportType
+    csv_text: str
+    source: str | None = None
+    delivery: DeliveryConfig | None = None
+
+
+class CSVImportError(BaseModel):
+    row: int
+    field: str | None = None
+    message: str
+
+
+class CSVImportWarning(BaseModel):
+    row: int
+    field: str | None = None
+    message: str
+
+
+class CSVImportResponse(BaseModel):
+    status: Literal["accepted", "partial", "rejected", "delivery_failed"]
+    import_type: CSVImportType
+    total: int
+    accepted: int
+    rejected: int
+    stored: int
+    posted: int
+    failed: int
+    source: str
+    delivery_mode: DestinationMode
+    delivery_attempts: int = 0
+    lot_codes: list[str]
+    errors: list[CSVImportError] = Field(default_factory=list)
+    warnings: list[CSVImportWarning] = Field(default_factory=list)
+    response: dict[str, Any] | None = None
+    error: str | None = None
