+++ app/schemas/simulation.py
@@ -1,0 +1,71 @@
+from __future__ import annotations
+
+from typing import Any, Literal
+
+from pydantic import BaseModel, Field, HttpUrl, field_validator
+
+from ..scenarios import ScenarioId
+from .domain import DestinationMode
+
+
+class DeliveryConfig(BaseModel):
+    mode: DestinationMode = DestinationMode.MOCK
+    endpoint: HttpUrl | None = None
+    api_key: str | None = None
+    tenant_id: str | None = None
+
+
+class SimulationConfig(BaseModel):
+    source: str = "codex-simulator"
+    scenario: ScenarioId = ScenarioId.LEAFY_GREENS_SUPPLIER
+    interval_seconds: float = 1.5
+    batch_size: int = 3
+    seed: int | None = 204
+    persist_path: str = "data/events.jsonl"
+    delivery: DeliveryConfig = Field(default_factory=DeliveryConfig)
+
+    @field_validator("interval_seconds")
+    @classmethod
+    def validate_interval(cls, value: float) -> float:
+        if value < 0:
+            raise ValueError("interval_seconds must be >= 0")
+        return value
+
+    @field_validator("batch_size")
+    @classmethod
+    def validate_batch_size(cls, value: int) -> int:
+        if value < 1 or value > 100:
+            raise ValueError("batch_size must be between 1 and 100")
+        return value
+
+
+class StartRequest(BaseModel):
+    config: SimulationConfig
+
+
+class StepRequest(BaseModel):
+    config: SimulationConfig | None = None
+
+
+class StatusResponse(BaseModel):
+    running: bool
+    config: SimulationConfig
+    stats: dict[str, Any]
+
+
+class StepResponse(BaseModel):
+    generated: int
+    posted: int
+    accepted: int = 0
+    rejected: int = 0
+    failed: int
+    lot_codes: list[str]
+    delivery_status: Literal["generated", "posted", "failed"]
+    delivery_mode: DestinationMode
+    delivery_attempts: int
+    response: dict[str, Any] | None = None
+    error: str | None = None
+
+
+class ResetResponse(BaseModel):
+    status: str
